api call from movieList and passing movie catagory as props
