+++ package.json
@@ -19,6 +19,7 @@
     "axios": "^1.11.0",
     "react": "19.1.0",
     "react-native": "0.80.2",
+    "react-native-fast-image": "^8.6.3",
     "react-native-safe-area-context": "^5.5.2",
     "react-native-screens": "^4.13.1",
     "react-native-vector-icons": "^10.3.0"
